rename: model에서 view로 패키지 수정
