+++ springbasic/src/main/java/com/programmers/springbasic/domain/voucher/view/VoucherInfo.java
@@ -8,7 +8,8 @@ public enum VoucherInfo {
     VOUCHER_VALUE_INFO_MESSAGE("value: "),
     VOUCHER_EXPIRE_INFO_MESSAGE("expireDate: "),
     VOUCHER_VALID_INFO_MESSAGE("valid: "),
-    VOUCHER_TYPE_INFO_MESSAGE("voucher type: ");
+    VOUCHER_TYPE_INFO_MESSAGE("voucher type: "),
+    VOUCHER_CUSTOMER_INFO_MESSAGE("customer ID: ");
 
     String infoMessage;
